Restructure spline functions, begin migration of code to implementation files
