+++ Doxyfile
@@ -917,19 +917,7 @@ RECURSIVE              = YES
 # Note that relative paths are relative to the directory from which doxygen is
 # run.
 
-EXCLUDE                = examples/README.md \
-                         include/smooth/external/lp2d.hpp \
-                         include/smooth/internal/bundle.hpp \
-                         include/smooth/internal/c1.hpp \
-                         include/smooth/internal/common.hpp \
-                         include/smooth/internal/lmpar.hpp \
-                         include/smooth/internal/lmpar_sparse.hpp \
-                         include/smooth/internal/macro.hpp \
-                         include/smooth/internal/se2.hpp \
-                         include/smooth/internal/se3.hpp \
-                         include/smooth/internal/so2.hpp \
-                         include/smooth/internal/so3.hpp \
-                         include/smooth/internal/tn.hpp
+EXCLUDE                = examples/README.md
 
 # The EXCLUDE_SYMLINKS tag can be used to select whether or not files or
 # directories that are symbolic links (a Unix file system feature) are excluded
@@ -945,7 +933,7 @@ EXCLUDE_SYMLINKS       = NO
 # Note that the wildcards are matched against the file with absolute path, so to
 # exclude all test directories for example use the pattern */test/*
 
-EXCLUDE_PATTERNS       =
+EXCLUDE_PATTERNS       = */detail/*
 
 # The EXCLUDE_SYMBOLS tag can be used to specify one or more symbol names
 # (namespaces, classes, functions, etc.) that should be excluded from the
