+++ include/smooth/detail/c1.hpp
