+++ include/smooth/detail/common.hpp
