+++ include/smooth/detail/lmpar.hpp
