+++ include/smooth/detail/macro.hpp
