+++ include/smooth/detail/se2.hpp
