+++ include/smooth/detail/se3.hpp
