+++ include/smooth/detail/so2.hpp
