+++ include/smooth/detail/so3.hpp
