+++ include/smooth/detail/utils.hpp
