+++ include/smooth/diff.hpp
@@ -34,7 +34,7 @@
 #include <Eigen/Core>
 #include <type_traits>
 
-#include "internal/utils.hpp"
+#include "detail/utils.hpp"
 #include "manifold.hpp"
 #include "wrt.hpp"
