+++ include/smooth/se2.hpp
@@ -31,9 +31,9 @@
 
 #include <complex>
 
-#include "internal/lie_group_base.hpp"
-#include "internal/macro.hpp"
-#include "internal/se2.hpp"
+#include "detail/lie_group_base.hpp"
+#include "detail/macro.hpp"
+#include "detail/se2.hpp"
 #include "lie_group.hpp"
 #include "map.hpp"
 #include "so2.hpp"
